Refactor server tests from use case to repository and create new tests
